Move random process-related functions to utils folder
