+++ autumn/tools/random_process/random_process.py
@@ -1,7 +1,9 @@
 from math import ceil
 from math import log, sqrt, pi
+from typing import Tuple
 
 from autumn.tools.curve import scale_up_function
+from .parameters import Time, RandomProcessParams
 
 
 class RandomProcess:
@@ -84,3 +86,52 @@ class RandomProcess:
 
         return log_likelihood
 
+
+class RandomProcessProc(ComputedValueProcessor):
+    """
+    Calculate the values of the random process
+    """
+
+    def __init__(self, rp_time_variant_func):
+        self.rp_time_variant_func = rp_time_variant_func
+
+    def process(self, compartment_values, computed_values, time):
+        return self.rp_time_variant_func(time)
+
+
+def set_up_random_process(start_time, end_time):
+    return RandomProcess(order=2, period=30, start_time=start_time, end_time=end_time)
+
+
+def get_random_process(
+        time_params: Time,
+        process_params: RandomProcessParams,
+        contact_rate_value: float,
+) -> Tuple[callable, callable]:
+    """
+    Work out the process that will contribute to the random process.
+
+    Args:
+        time_params: Start and end time of the model
+        process_params: Parameters relating to the random process
+        contact_rate_value: The risk of transmission per contact
+
+    Returns:
+        The random process function and the contact rate (here a summer-ready format transition function)
+
+    """
+
+    # Build the random process, using default values and coefficients
+    rp = set_up_random_process(time_params.start, time_params.end)
+
+    # Update random process details based on the model parameters
+    rp.update_config_from_params(process_params)
+
+    # Create function returning exp(W), where W is the random process
+    rp_time_variant_func = rp.create_random_process_function(transform_func=lambda w: exp(w))
+
+    # Create the time-variant contact rate that uses our computed random process
+    def contact_rate_func(t, computed_values):
+        return contact_rate_value * computed_values["transformed_random_process"]
+
+    return rp_time_variant_func, contact_rate_func
